Better plotting, filtering, and new fitting functions
Plotting: Peak labels had a tendency to overlap. Experimented with different ways of plotting peaks including turning them on their side, tilting them at an angle, changing what side of the peak they were plotted on, and changing them to integer values. I found that plotting integer values along with centering the label above the peak worked best.

Filtering: I created a new filtering algorithm. Essentially I found all of the peaks by looking only 1.005* the mean of all the points in this part of the spectrum, and this provided all of the peaks, for HCl-35 and HCl-37. I then checked all the peaks to see if they were smaller than the peaks on either side of them.

Fitting: Created functions that will fit the values to the quantum numbers associated with their transition. The functions vary depending on P-Branch or R-Branch, and whether or not they are the fundamental branch or just the first overtone.