Fixed Algorithm After New Test Data
Discard peaks that are insanely small compared to neighbors.
Note that first peak in branch might be smaller than both peaks on either side
Correct algorithm to always keep first P-Branch Peak.
Correct algorithm to compare prominance of first R-Branch Peak with first P-Branch Peak. If they are similar, it is the true R branch Peak. If they are not similar, it is a Cl-37 Peak, and the next peak to the left should be used.